fix(kernel): fix second uint16_t narrowing missed by different indentation
The replace_all only matched the 8-space-indented occurrence (line 345)
but missed the 4-space-indented one (line 394) due to exact string match.

+++ vllm_ascend/attention/ops/tq_fused_decode/op_kernel/tq_fused_decode_kernel.cpp
@@ -391,7 +391,7 @@ __aicore__ inline void KernelTqFusedDecode::Process() {
     GlobalTensor<half> outOffsetGm;
     outOffsetGm.SetGlobalBuffer(reinterpret_cast<__gm__ half*>(outputBase_) + outOff);
     DataCopyPad(outOffsetGm, outFp16,
-                {1, static_cast<uint32_t>(tiling.headDim * sizeof(half)), 0, 0});
+                {1, static_cast<uint16_t>(tiling.headDim * sizeof(half)), 0, 0});
     pipe_barrier(PIPE_V);
     slotQueue.FreeTensor(slotLocal);
 }
